Telegram File Sender/requirements.txt
aiogram
aiofiles
chardet
python-dotenv
